heracles-api/requirements.txt
# Heracles API Dependencies
# =========================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# LDAP
ldap3==2.9.1

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
orjson==3.9.10

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.0
ruff==0.1.13
mypy==1.8.0
